requirements.txt
Django==2.2
django-tinymce4-lite==1.7.5
jsmin==2.2.2
mysqlclient==1.4.2.post1
Pillow==6.0.0
pytz==2019.1
sqlparse==0.3.0
